create vibration builder
